flask/requirements.txt
Flask
mysql-connector-python
pyodbc
oracledb
psycopg2-binary
